Cleanup linfcn.h
Replace long with int32_t, added named parameters to functions.

+++ src/Libraries/2D/Source/Flat8/fl8hlin.c
@@ -65,7 +65,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 /* clut should be in _ns everywhere, it doesn't need its own function */
 
-void gri_flat8_uhline_ns(short x0, short y0, short x1, long c, long parm) {
+void gri_flat8_uhline_ns(short x0, short y0, short x1, int32_t c, int32_t parm) {
     uchar *p;
     short t;
 
@@ -80,7 +80,7 @@ void gri_flat8_uhline_ns(short x0, short y0, short x1, long c, long parm) {
     memset(p, c, x1 - x0 + 1);
 }
 
-void gri_flat8_uhline_clut(short x0, short y0, short x1, long c, long parm) {
+void gri_flat8_uhline_clut(short x0, short y0, short x1, int32_t c, int32_t parm) {
     uchar *p;
     short t;
 
@@ -90,12 +90,12 @@ void gri_flat8_uhline_clut(short x0, short y0, short x1, long c, long parm) {
         x1 = t;
     }
 
-    c = (long)(((uchar *)parm)[c]);
+    c = (int32_t)(((uchar *)parm)[c]);
     p = grd_bm.bits + y0 * grd_bm.row + x0;
     memset(p, c, x1 - x0 + 1);
 }
 
-void gri_flat8_uhline_xor(short x0, short y0, short x1, long c, long parm) {
+void gri_flat8_uhline_xor(short x0, short y0, short x1, int32_t c, int32_t parm) {
     uchar *p;
     short t;
 
@@ -110,7 +110,7 @@ void gri_flat8_uhline_xor(short x0, short y0, short x1, long c, long parm) {
 }
 
 /* punt */
-void gri_flat8_uhline_blend(short x0, short y0, short x1, long c, long parm) {
+void gri_flat8_uhline_blend(short x0, short y0, short x1, int32_t c, int32_t parm) {
     uchar *p;
     short t;
