+++ src/Libraries/2D/Source/Flat8/fl8slin.c
@@ -58,17 +58,17 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
         p[x] = i;             \
     } while (0)
 
-void gri_flat8_usline_norm(long c, long parm, grs_vertex *v0, grs_vertex *v1) {
+void gri_flat8_usline_norm(int32_t c, int32_t parm, grs_vertex *v0, grs_vertex *v1) {
 #include "fl8slin.h"
 }
 
 #undef macro_plot_i
 #define macro_plot_i(x, p, i)              \
     do {                                   \
-        p[x] = (long)(((uchar *)parm)[i]); \
+        p[x] = (int32_t)(((uchar *)parm)[i]); \
     } while (0)
 
-void gri_flat8_usline_clut(long c, long parm, grs_vertex *v0, grs_vertex *v1) {
+void gri_flat8_usline_clut(int32_t c, int32_t parm, grs_vertex *v0, grs_vertex *v1) {
 #include "fl8slin.h"
 }
 
@@ -78,7 +78,7 @@ void gri_flat8_usline_clut(long c, long parm, grs_vertex *v0, grs_vertex *v1) {
         p[x] = p[x] ^ i;      \
     } while (0)
 
-void gri_flat8_usline_xor(long c, long parm, grs_vertex *v0, grs_vertex *v1) {
+void gri_flat8_usline_xor(int32_t c, int32_t parm, grs_vertex *v0, grs_vertex *v1) {
 #include "fl8slin.h"
 }
 
@@ -90,6 +90,6 @@ void gri_flat8_usline_xor(long c, long parm, grs_vertex *v0, grs_vertex *v1) {
         p[x] = i;             \
     } while (0)
 
-void gri_flat8_usline_blend(long c, long parm, grs_vertex *v0, grs_vertex *v1) {
+void gri_flat8_usline_blend(int32_t c, int32_t parm, grs_vertex *v0, grs_vertex *v1) {
 #include "fl8slin.h"
 }
