+++ src/Libraries/2D/Source/Flat8/fl8wclin.c
@@ -54,7 +54,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
         } while (x > x_new)
 
 // MLA #pragma off (unreferenced)
-void gri_flat8_wire_poly_ucline_norm(long c, long parm, grs_vertex *v0, grs_vertex *v1) {
+void gri_flat8_wire_poly_ucline_norm(int32_t c, int32_t parm, grs_vertex *v0, grs_vertex *v1) {
 #include "fl8wclin.h"
 }
     // MLA #pragma on (unreferenced)
@@ -81,7 +81,7 @@ void gri_flat8_wire_poly_ucline_norm(long c, long parm, grs_vertex *v0, grs_vert
             r0 += dr, g0 += dg, b0 += db;                \
         } while (x > x_new)
 
-void gri_flat8_wire_poly_ucline_clut(long c, long parm, grs_vertex *v0, grs_vertex *v1) {
+void gri_flat8_wire_poly_ucline_clut(int32_t c, int32_t parm, grs_vertex *v0, grs_vertex *v1) {
     uchar *clut = (uchar *)parm;
 #include "fl8wclin.h"
 }
